Refactored code and UX
