--- src/features/P2pTradeManager/index.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import TaskList from '../P2PTaskList';
-
-const P2pTradeManager = () => {
-    return (
-        <div>
-            <TaskList />
-        </div>
-    );
-};
-
-export default P2pTradeManager;
